requirements.txt
transformers
torch
pillow
requests
beautifulsoup4
pdfplumber
